Changing listingData size type to float.

+++ src/redux/ducks/listing.js
@@ -412,7 +412,7 @@ const mutationUpdate = gql`
     $currency: String
     $isAbsorvedFee: Boolean
     $capacity: Int
-    $size: Int
+    $size: Float
     $meetingRooms: Int
     $isFurnished: Boolean
     $carSpace: Int
